# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AURELION - Luxury Aruba Adventure Planning Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Activity {
  id: number;
  title: string;
  category: string;
  difficulty: string;
  durationMinutes: number;
  priceLow: number;
  priceHigh: number;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  reviewSummary?: string | null;
  tags: string[];
  description: string;
  /** @nullable */
  whatToBring?: string | null;
  /** @nullable */
  whatToExpect?: string | null;
  createdAt: string;
}

export interface ActivityDetail {
  id: number;
  title: string;
  category: string;
  difficulty: string;
  durationMinutes: number;
  priceLow: number;
  priceHigh: number;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  reviewSummary?: string | null;
  tags: string[];
  description: string;
  /** @nullable */
  whatToBring?: string | null;
  /** @nullable */
  whatToExpect?: string | null;
  /** @nullable */
  basicBookingGuide?: string | null;
  /** @nullable */
  providerName?: string | null;
  /** @nullable */
  providerWebsite?: string | null;
  /** @nullable */
  providerPhone?: string | null;
  createdAt: string;
}

export interface CategorySummary {
  category: string;
  count: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface UserSession {
  id?: number;
  name?: string;
  email?: string;
  role?: string;
  tier?: string;
  isAuthenticated: boolean;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface Itinerary {
  id: number;
  userId: number;
  title: string;
  totalDays: number;
  tierType: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface ItineraryItemWithActivity {
  id: number;
  itineraryId: number;
  activityId: number;
  dayNumber: number;
  timeSlot: string;
  /** @nullable */
  notes?: string | null;
  activity: Activity;
}

export interface ItineraryDetail {
  id: number;
  userId: number;
  title: string;
  totalDays: number;
  tierType: string;
  status: string;
  createdAt: string;
  updatedAt: string;
  items: ItineraryItemWithActivity[];
}

export interface ItineraryItem {
  id: number;
  itineraryId: number;
  activityId: number;
  dayNumber: number;
  timeSlot: string;
  /** @nullable */
  notes?: string | null;
}

export interface CreateItineraryBody {
  title: string;
  totalDays: number;
}

export interface UpdateItineraryBody {
  title?: string;
  totalDays?: number;
  status?: string;
}

export interface AddItineraryItemBody {
  activityId: number;
  dayNumber: number;
  timeSlot: string;
  notes?: string;
}

export interface UpdateItineraryItemBody {
  dayNumber?: number;
  timeSlot?: string;
  /** @nullable */
  notes?: string | null;
}

export interface ItineraryExport {
  itinerary: ItineraryDetail;
  tierType: string;
  exportedAt: string;
}

export interface ChatSession {
  id: number;
  userId: number;
  /** @nullable */
  itineraryId?: number | null;
  createdAt: string;
}

export interface CreateChatSessionBody {
  itineraryId?: number;
}

export interface ChatMessage {
  id: number;
  sessionId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface SendMessageBody {
  content: string;
}

export interface Purchase {
  id: number;
  userId: number;
  /** @nullable */
  itineraryId?: number | null;
  productType: string;
  amount: number;
  status: string;
  createdAt: string;
}

export interface CreateCheckoutBody {
  itineraryId: number;
  productType: string;
}

export interface CheckoutResponse {
  url: string;
}

export interface AdminCreateActivityBody {
  title: string;
  category: string;
  difficulty: string;
  description: string;
  durationMinutes: number;
  priceLow: number;
  priceHigh: number;
  location: string;
  imageUrl?: string;
  reviewSummary?: string;
  whatToBring?: string;
  whatToExpect?: string;
  basicBookingGuide?: string;
  premiumBookingGuide?: string;
  insiderTips?: string;
  warnings?: string;
  bestTimeOfDay?: string;
  tags?: string[];
  providerName?: string;
  providerWebsite?: string;
  providerPhone?: string;
  providerEmail?: string;
  providerWhatsapp?: string;
}

export interface IngestUrlBody {
  url: string;
}

export interface ScrapedActivityPreview {
  title?: string;
  description?: string;
  url: string;
  suggestedCategory?: string;
  rawText: string;
}

export interface DashboardSummary {
  totalItineraries: number;
  totalActivities: number;
  recentItineraries: Itinerary[];
  activitiesByCategory: CategorySummary[];
  hasPremium: boolean;
}

export type ListActivitiesParams = {
  category?: string;
  difficulty?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  maxDuration?: number;
  location?: string;
};