# lib/api-spec/openapi.yaml
...
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AURELION - Luxury Aruba Adventure Planning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List all activities
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: maxDuration
          in: query
          schema:
            type: number
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
...
  /activities/{id}:
    get:
      operationId: getActivity
      tags: [activities]
      summary: Get single activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityDetail"
        "404":
          description: Not found
...
  /admin/activities:
    post:
      operationId: adminCreateActivity
      tags: [admin]
      summary: Create activity (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateActivityBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /admin/activities/{id}:
    patch:
      operationId: adminUpdateActivity
      tags: [admin]
      summary: Update activity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateActivityBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: adminDeleteActivity
      tags: [admin]
      summary: Delete activity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/ingest:
    post:
      operationId: adminIngestUrl
      tags: [admin]
      summary: Ingest activity from URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestUrlBody"
      responses:
        "200":
          description: Scraped activity preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapedActivityPreview"
...
components:
  schemas:
...
    Activity:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        difficulty:
          type: string
        durationMinutes:
          type: integer
        priceLow:
          type: number
        priceHigh:
          type: number
        location:
          type: string
        imageUrl:
          type: ["string", "null"]
        reviewSummary:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        description:
          type: string
        whatToBring:
          type: ["string", "null"]
        whatToExpect:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - category
        - difficulty
        - durationMinutes
        - priceLow
        - priceHigh
        - location
        - description
        - tags
        - createdAt
    ActivityDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        difficulty:
          type: string
        durationMinutes:
          type: integer
        priceLow:
          type: number
        priceHigh:
          type: number
        location:
          type: string
        imageUrl:
          type: ["string", "null"]
        reviewSummary:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        description:
          type: string
        whatToBring:
          type: ["string", "null"]
        whatToExpect:
          type: ["string", "null"]
        basicBookingGuide:
          type: ["string", "null"]
        providerName:
          type: ["string", "null"]
        providerWebsite:
          type: ["string", "null"]
        providerPhone:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - category
        - difficulty
        - durationMinutes
        - priceLow
        - priceHigh
        - location
        - description
        - tags
        - createdAt
...
    AdminCreateActivityBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        difficulty:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        priceLow:
          type: number
        priceHigh:
          type: number
        location:
          type: string
        imageUrl:
          type: string
        reviewSummary:
          type: string
        whatToBring:
          type: string
        whatToExpect:
          type: string
        basicBookingGuide:
          type: string
        premiumBookingGuide:
          type: string
        insiderTips:
          type: string
        warnings:
          type: string
        bestTimeOfDay:
          type: string
        tags:
          type: array
          items:
            type: string
        providerName:
          type: string
        providerWebsite:
          type: string
        providerPhone:
          type: string
        providerEmail:
          type: string
        providerWhatsapp:
          type: string
      required:
        - title
        - category
        - difficulty
        - description
        - durationMinutes
        - priceLow
        - priceHigh
        - location
    IngestUrlBody:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    ScrapedActivityPreview:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        url:
          type: string
        suggestedCategory:
          type: string
        rawText:
          type: string
      required:
        - url
        - rawText
...